ELMU #136 Use injection to instantiate the user request handler
